tests: refactor service tests to use mock module
